requirements.txt.txt
pytest
allure-pytest
faker
pydantic
requests
